Remove redundant casts

+++ ImageScaler/src/me/roan/imagescaler/Worker.java
@@ -116,7 +116,7 @@ public class Worker{
 			ImageReader reader = readers.next();
 			reader.setInput(ImageIO.createImageInputStream(file));
 			BufferedImage img = reader.read(0);
-			BufferedImage output = new ResampleOp((int)Math.round((double)img.getWidth() * Main.scale), (int)Math.round((double)img.getHeight() * Main.scale), Main.mode.mode).filter(img, null);
+			BufferedImage output = new ResampleOp((int)Math.round(img.getWidth() * Main.scale), (int)Math.round(img.getHeight() * Main.scale), Main.mode.mode).filter(img, null);
 			out.createNewFile();
 			Iterator<ImageWriter> writers = ImageIO.getImageWritersBySuffix(ext);
 			if(!writers.hasNext()){
